requirements.txt
torch>=2.1
torchaudio
transformers>=4.40
datasets>=2.18
peft>=0.10
scikit-learn
matplotlib
seaborn
soundfile
pyyaml
tqdm
ipython
jupyter
